requirements.txt
asgiref==3.2.6
Django==3.0.4
django-webpack-loader==0.7.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
